Move various helpers into API project
There were a number of classes in DLCS.Model and DLCS.Repository that
will only be used in API so moved over. Related to AssetQuery model,
incoming commit will refactor some further.
