+++ src/protagonist/API/Features/Customer/Requests/GetQueriedAllImages.cs
@@ -1,6 +1,7 @@
-﻿using API.Infrastructure.Requests;
+﻿using API.Features.Assets.Query;
+using API.Infrastructure.Page;
+using API.Infrastructure.Requests;
 using DLCS.Model.Assets;
-using DLCS.Model.Page;
 using DLCS.Repository;
 using DLCS.Repository.Assets;
 using MediatR;
