+++ src/protagonist/API/Features/DeliveryChannels/Requests/DefaultDeliveryChannels/GetDefaultDeliveryChannels.cs
@@ -1,6 +1,6 @@
-﻿using API.Infrastructure.Requests;
+﻿using API.Infrastructure.Page;
+using API.Infrastructure.Requests;
 using DLCS.Model.DeliveryChannels;
-using DLCS.Model.Page;
 using DLCS.Repository;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
