+++ src/protagonist/API/Features/Image/ImagesController.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using API.Converters;
 using API.Exceptions;
+using API.Features.Assets.Query;
 using API.Features.Image.Requests;
 using API.Features.Image.Validation;
 using API.Features.Space.Requests;
