+++ src/protagonist/API/Features/Queues/Requests/GetBatches.cs
@@ -1,6 +1,6 @@
-﻿using API.Infrastructure.Requests;
+﻿using API.Infrastructure.Page;
+using API.Infrastructure.Requests;
 using DLCS.Model.Assets;
-using DLCS.Model.Page;
 using DLCS.Repository;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
