+++ src/protagonist/API/Infrastructure/Page/PagingHelpers.cs
@@ -1,11 +1,7 @@
-﻿using System;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
-using DLCS.Model.Page;
+﻿using API.Infrastructure.Requests;
 using Microsoft.EntityFrameworkCore;
 
-namespace DLCS.Repository;
+namespace API.Infrastructure.Page;
 
 public static class PagingHelpers
 {
@@ -30,36 +26,28 @@ public static class PagingHelpers
         Func<IQueryable<T>, IQueryable<T>>? entityOperations = null,
         CancellationToken cancellationToken = default) where T : class
     {
-        try
+        var workingEntities = filter(entities);
+        var entityQuery = workingEntities;
+        if (entityOperations != null)
         {
-            var workingEntities = filter(entities);
-            var entityQuery = workingEntities;
-            if (entityOperations != null)
-            {
-                entityQuery = entityOperations(entityQuery);
-            }
-
-            var result = new PageOf<T>
-            {
-                Page = request.Page,
-                Total = await workingEntities.CountAsync(cancellationToken: cancellationToken),
-                Entities = await entityQuery.WithPaging(request).ToListAsync(cancellationToken),
-                PageSize = request.PageSize
-            };
-
-            return result;
+            entityQuery = entityOperations(entityQuery);
         }
-        catch (Exception ex)
+
+        var result = new PageOf<T>
         {
-            var x = ex.Message;
-            throw;
-        }
+            Page = request.Page,
+            Total = await workingEntities.CountAsync(cancellationToken: cancellationToken),
+            Entities = await entityQuery.WithPaging(request).ToListAsync(cancellationToken),
+            PageSize = request.PageSize
+        };
+
+        return result;
     }
 
     /// <summary>
     /// Apply Skip/Take logic on IQueryable using paged request
     /// </summary>
-    public static IQueryable<T> WithPaging<T>(this IQueryable<T> query, IPagedRequest request)
+    private static IQueryable<T> WithPaging<T>(this IQueryable<T> query, IPagedRequest request)
         => query
             .Skip((request.Page - 1) * request.PageSize)
             .Take(request.PageSize);
