+++ src/protagonist/API/Infrastructure/Requests/IAssetFilterableRequest.cs
@@ -1,0 +1,11 @@
+using API.Features.Assets.Query;
+
+namespace API.Infrastructure.Requests;
+
+/// <summary>
+/// Marker interface for requests that can be filtered by Asset Query Syntax
+/// </summary>
+public interface IAssetFilterableRequest
+{
+    AssetQueryModel AssetQueryModel { get; }
+}
