+++ src/protagonist/API/Infrastructure/Requests/IPagedRequest.cs
@@ -1,0 +1,10 @@
+namespace API.Infrastructure.Requests;
+
+/// <summary>
+/// Represents a request for a paged resource
+/// </summary>
+public interface IPagedRequest
+{
+    int Page { get; set; }
+    int PageSize { get; set; }
+}
